Clarify and correct readlines/writelines examples
Show that Python's fileobj.readlines() returns lines with line terminators, while vim's readlines() returns lines without the terminators.

Correct the writelines example: the code as written passed a string instead of a list of strings, which actually works, but very inefficiently (writing one character at a time).  A better way would be to `fh.write('\n'.join(lines))`, but that is still incorrect: it omits the trailing newline which is mandatory/customary on text files in Unix-like OSes.  Instead show how to print multiple lines using the print() function in Python 3 if you need the newlines to be added, and show how to use writelines() if the newlines are already there.